db manager changes
DbManager is now a bean which can be injected. local websphere testing
with database has changed, now requirers config of local server to set
data source. This means we can use the exact same code for both blumix
release and local tesing. also db2jcc lib can be removed since this i
managed by server now in both cases => we also get connection pooling
since it is enabled by default.

Example websphere config files has ben put in dir
"websphere_server_config".
