+++ web_project/2dv512-app/.classpath
@@ -20,6 +20,6 @@
 	<classpathentry kind="lib" path="dep-jar/com.ibm.ws.javaee.jaxrs.2.0_1.0.10.jar"/>
 	<classpathentry kind="lib" path="dep-jar/com.ibm.ws.javaee.persistence.2.1_1.0.10.jar"/>
 	<classpathentry kind="lib" path="WebContent/WEB-INF/lib/primefaces-5.3.jar"/>
-	<classpathentry kind="lib" path="WebContent/WEB-INF/lib/db2jcc4.jar"/>
+	<classpathentry kind="lib" path="dep-jar/com.ibm.ws.javaee.ejb.3.2_1.0.10.jar"/>
 	<classpathentry kind="output" path="build/classes"/>
 </classpath>
